bitchat-tui: skip test on linux ci

+++ Formula/b/bitchat-tui.rb
@@ -13,12 +13,17 @@ class BitchatTui < Formula
   end
 
   test do
-    output_log = testpath/"output.log"
-    pid = spawn bin/"bitchat-tui", [:out, :err] => output_log.to_s
-    sleep 1
-    assert_match "Public Chat", output_log.read
-  ensure
-    Process.kill("TERM", pid)
-    Process.wait(pid)
+    # failed with Linux CI, `No such device or address` error
+    return if OS.linux? && ENV["HOMEBREW_GITHUB_ACTIONS"]
+
+    begin
+      output_log = testpath/"output.log"
+      pid = spawn bin/"bitchat-tui", [:out, :err] => output_log.to_s
+      sleep 1
+      assert_match "Public Chat", output_log.read
+    ensure
+      Process.kill("TERM", pid)
+      Process.wait(pid)
+    end
   end
 end
